test(condo): DOMA-6931 updated tests
