+++ apps/condo/domains/news/schema/NewsItem.test.js
@@ -31,17 +31,6 @@ const {
     createTestNewsItemScope,
 } = require('@condo/domains/news/utils/testSchema')
 const {
-    NEWS_ITEM_COMMON_MESSAGE_TYPE,
-    DEVICE_PLATFORM_ANDROID,
-    APP_RESIDENT_ID_ANDROID,
-    MESSAGE_SENT_STATUS,
-} = require('@condo/domains/notification/constants/constants')
-const { Message, syncRemoteClientByTestClient } = require('@condo/domains/notification/utils/testSchema')
-const {
-    getRandomTokenData,
-    getRandomFakeSuccessToken,
-} = require('@condo/domains/notification/utils/testSchema/helpers')
-const {
     createTestOrganizationEmployeeRole,
     createTestOrganizationEmployee,
     createTestOrganization,
@@ -828,21 +817,6 @@ describe('NewsItems', () => {
             )
         })
 
-        // todo(doma-6931): rewrite test
-        test.skip('must throw an error on trying to edit the news item which already been sent', async () => {
-            const [sentNewsItem] = await createTestNewsItem(adminClient, dummyO10n, { sentAt: dayjs().toISOString() })
-            await expectToThrowGQLError(
-                async () => await updateTestNewsItem(adminClient, sentNewsItem.id, { title: faker.lorem.words(3) }),
-                {
-                    code: 'BAD_USER_INPUT',
-                    type: 'EDIT_DENIED_ALREADY_SENT',
-                    message: 'The sent news item is restricted from editing',
-                    mutation: 'updateNewsItem',
-                    messageForUser: 'api.newsItem.EDIT_DENIED_ALREADY_SENT',
-                },
-            )
-        })
-
         test('must throw an error on trying to edit the published news item', async () => {
             const [sentNewsItem] = await createTestNewsItem(adminClient, dummyO10n)
             await createTestNewsItemScope(adminClient, sentNewsItem)
@@ -999,114 +973,6 @@ describe('NewsItems', () => {
                 const newsItems3 = await NewsItem.getAll(residentClient1, {})
                 expect(newsItems3).toHaveLength(1)
             }, { delay: 4 * 1000 })
-        })
-    })
-
-    describe('notifications', () => {
-        // TODO(DOMA-6931): not work without task.delay in afterChange
-        test.skip('the user receives a push notification on a news item created and does not receive notification for 2nd news item', async () => {
-            const residentClient1 = await makeClientWithResidentUser()
-            const [o10n] = await createTestOrganization(adminClient)
-            const [property] = await createTestProperty(adminClient, o10n)
-
-            const unitType1 = FLAT_UNIT_TYPE
-            const unitName1 = faker.lorem.word()
-
-            const [resident] = await createTestResident(adminClient, residentClient1.user, property, {
-                unitType: unitType1,
-                unitName: unitName1,
-            })
-
-            // News item for particular unit
-            const [newsItem1] = await createTestNewsItem(
-                adminClient,
-                o10n,
-                {
-                    title: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
-                    body: 'Commodo viverra maecenas accumsan lacus vel facilisis volutpat est velit. Et malesuada fames ac turpis egestas sed tempus urna et. At augue eget arcu dictum varius duis at. Tempus quam pellentesque nec nam aliquam sem et tortor consequat. Enim sit amet venenatis urna cursus eget nunc scelerisque viverra. Urna cursus eget nunc scelerisque viverra. Ornare aenean euismod elementum nisi quis eleifend quam. Quis hendrerit dolor magna eget est. Gravida cum sociis natoque penatibus et.',
-                })
-            await createTestNewsItemScope(adminClient, newsItem1, {
-                property: { connect: { id: property.id } },
-                unitType: unitType1,
-                unitName: unitName1,
-            })
-
-            const payload = getRandomTokenData({
-                devicePlatform: DEVICE_PLATFORM_ANDROID,
-                appId: APP_RESIDENT_ID_ANDROID,
-                pushToken: getRandomFakeSuccessToken(),
-            })
-
-            await syncRemoteClientByTestClient(residentClient1, payload)
-
-            const messageWhere = { user: { id: residentClient1.user.id }, type: NEWS_ITEM_COMMON_MESSAGE_TYPE }
-
-            // Publish news item to make it send-able
-            const [updatedItem1] = await publishTestNewsItem(adminClient, newsItem1.id)
-
-            await waitFor(async () => {
-                const messages = await Message.getAll(adminClient, messageWhere)
-
-                expect(messages).toBeDefined()
-                expect(messages).toHaveLength(1)
-
-                const message1 = messages[0]
-
-                expect(message1).toBeDefined()
-                expect(message1.id).toMatch(UUID_RE)
-
-                expect(message1).toEqual(expect.objectContaining({
-                    status: MESSAGE_SENT_STATUS,
-                    processingMeta: expect.objectContaining({
-                        // old way check
-                        transport: 'push',
-
-                        // ADR-7 way check
-                        transportsMeta: [expect.objectContaining({
-                            transport: 'push',
-                        })],
-                    }),
-                    meta: expect.objectContaining({
-                        title: 'Lorem ipsum dolor sit amet, consectetur...',
-                        body: 'Commodo viverra maecenas accumsan lacus vel facilisis volutpat est velit. Et malesuada fames ac turpis egestas sed tempus urna et. At augue eget...',
-                        data: expect.objectContaining({
-                            newsItemId: newsItem1.id,
-                            residentId: resident.id,
-                            userId: residentClient1.user.id,
-                            userRelatedResidentsIds: resident.id,
-                            organizationId: o10n.id,
-                            validBefore: null,
-                            dateCreated: updatedItem1.publishedAt,
-                        }),
-                    }),
-                }))
-            }, { delay: (SENDING_DELAY_SEC + 3) * 1000 })
-
-            // This news item shouldn't send notification for the same user
-            const [newsItem2] = await createTestNewsItem(adminClient, o10n)
-            await createTestNewsItemScope(adminClient, newsItem2, {
-                property: { connect: { id: property.id } },
-            })
-
-            // Publish 2nd news item...
-            await publishTestNewsItem(adminClient, newsItem2.id)
-            //... and shouldn't see any sent message for it
-            await waitFor(async () => {
-                const messages = await Message.getAll(adminClient, messageWhere)
-
-                expect(messages).toHaveLength(2)
-                expect(messages).toEqual(expect.arrayContaining([
-                    expect.objectContaining({ type: 'NEWS_ITEM_COMMON_MESSAGE_TYPE', status: 'sent' }),
-                    expect.objectContaining({
-                        type: 'NEWS_ITEM_COMMON_MESSAGE_TYPE',
-                        status: 'throttled',
-                        meta: expect.objectContaining({
-                            data: expect.objectContaining({ newsItemId: newsItem2.id }),
-                        }),
-                        processingMeta: expect.objectContaining({ error: expect.stringContaining('1 message per 3600 sec for user. The latest message was at ') }),
-                    }),
-                ]))
-            }, { delay: (SENDING_DELAY_SEC + 3) * 1000 })
         })
     })
